+++ MyPlotItem.h
@@ -58,7 +58,7 @@ public:
 		return isRealTimeCurve_;
 	}
 
-	virtual std::optional<double> value(int32_t const x) const noexcept = 0;
+	virtual std::optional<double> value(double const x) const noexcept = 0;
 public slots:
 	virtual void handleData(std::pair<std::vector<double>, std::vector<Timestamp6991>> const& data) = 0;
 signals:
